add move enqueuing
This commit should allow for moves to be enqueued. At
this point, we check that the provided game exists,
the request body provides a valid user id, and a list
of moves. Move validation does not work yet. Each game
is hardcoded to include a "Foo" and "Bar" player for now