add basic game clock
This commit introduces an internal "clock" for each game,
starting once the second player has joined. For now, it
will simply log to the console that players are moving,
and kill the timer when a new game is created. But the
commit lays most of the groundwork for the clock
architecture. Moves are pulled off each player's move
queue, converted to actions (hard coded to be just a
position change for now), then executed in order.